requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
openai==0.11.4
python-dotenv==0.19.2
databases==0.5.5
sentence-transformers==2.1.0
matplotlib==3.4.3
openai-whisper==1.0.0
ffmpeg-python==0.2.0
librosa==0.8.1
transformers==4.12.0
